[ADD] wsgi server

+++ backend/database/__init__.py
@@ -10,6 +10,10 @@ from exts import db
 from .models import *
 
 def Initialize():
+    user = User.query.filter_by(username="admin").first()
+    if user is not None:
+        return
+    
     db.drop_all()
     db.create_all()
     # user
